Redirect from unversioned readme URLs to new 2.18.x (#35)
This duplicates all the 2.18 readme redirect URLs, except without the
`/v2.18/` bit. This is because readme supports these unversioned URLs,
and in fact visiting the default website uses these unversioned version:
thus, most URLs people have shared will be the unversioned ones (and
even visiting an explicitly versioned link with the current main version
will redirect to these unversioned ones).

Fixes #5 (taking approach 2).

NB. if 2.19.0 is released before the switch-over, we'll want to revert
this and redo it for 2.19. Process: find the segment with the versioned
URLs with the version of interest, duplicate it and do a search replace
to strip out `/v2.NN/`.