Check reference docs are up to date in CI (#64)
This does a basic version of validating that the checked-in reference
docs are valid, i.e. regenerating them gives the same result. It does
this by just running the generation and checking that there's no
changes.

This ensures that we don't, for instance, make a change to the templates
and then forget to regenerate everything.