Collapse "overview" pages (#109)
This change simply "collapses" `/foo/foo-overview` paths to be `/foo` by
renaming the `-overview.mdx` files to be `index.mdx` (in most cases. For
Python and plugins, there was a different rename since their structure
wasn't the same)

This also has the nice effect of making the sidebar dropdown for each
section a clickable link to the section itself.