Sort fields and options before generating reference (#140)
Noticed in #139 that fields and options were jumping around with no
reason. After some testing, I've concluded that the JSON output from
`pants help-all` is non-deterministically ordered. Since the JS just
loads that data, it ends up shifting around the docs as well.

I had a quick peek in the help-all code and while I think it would be
possible to affect it there, this seems easier/quicker to fix (and only
pays the price here). The downside is that this is a bigger diff than
the spurious changes that now occur.

I did a second commit where I played around with also sorting required
fields before optional fields, which is what the CLI seems to do. If we
do that, we should probably also annotate required fields in some way. I
think there's pros and cons with both.