Add PR CI check that humans don't edit auto-generated files (#104)
This fixes #76 by validating that only PRs that have the
`automation:sync-docs` label make changes to the files that are directly
from the pantsbuild/pants repo, in the `docs/` and `versioned_docs/*`
directories:

- prose docs in `docs/`
- help JSON in `reference/help-all.json`

Generally speaking, a human shouldn't change these files directly. The
sync docs automation attaches the label, and if a human knows what
they're doing, they can add the label themselves.

This is an "MVP" approach, checking if there's changes and outputting a
vaguely explanatory message. Future iterations might comment directly on
the PR and/or construct links about which file(s) need to be updated,
but that's a bit fiddly.

This PR also updates the error about the reference docs, in case someone
edits those files directly too.

Examples: 

- error when there's an inappropriate change:
https://github.com/pantsbuild/pantsbuild.org/actions/runs/7534490860/job/20508964554
- no error when there's a change with the `automation:sync-docs` label:
https://github.com/pantsbuild/pantsbuild.org/actions/runs/7534631863/job/20509314384
- (regression test) still an error when the reference auto-generated
files are manually changed:
https://github.com/pantsbuild/pantsbuild.org/actions/runs/7534496644/job/20508980096