Upload build from PR CI for easier previewing (#150)
This has PR CI upload the resulting static site output, once the build
finishes.

This is a stop gap for
https://github.com/pantsbuild/pantsbuild.org/issues/28 that's seems like
it's better than nothing: to preview a PR, click through to the build
summary, download the artifact and take a look. This saves a reviewer
from needing to check-out the branch and build it, just do a (100MB)
download instead.