Check for .0a0 for new version creation, not .a0 (#163)
Minor typo in the check for alpha versions meant it didn't trigger: it's
intending to search for `2.x.0a0`, i.e. first alpha version leading up
to the first stable release in `2.x` series... but missed the 0 between
the `.` and the `a`.

This explains
(https://github.com/pantsbuild/pantsbuild.org/pull/162#issuecomment-1968056882):
why we don't have a `versioned_docs/version-2.20` directory despite #153
landing.

This PR doesn't create that directory, just fixing the bug in the
automation. I've retriggered the sync from this PR's branch to do that:
#164