Centralise versioning configuration (#176)
The goal here is to centralise the logic around versions, creating a
list of all the details about each version (e.g. full version number,
whether it's a pre-release or unmaintained etc.), that downstream
processing can reference directly rather than recompute. This, for
instance, eliminates the duplicated logic around which versions are
maintained.

The goal here is to make changes like #174 (and maybe #159) easier. This
PR doesn't/shouldn't change observable behaviour.

This also removes the explicit `disableVersioning` flag, since that
seems to require changing the contents of the `versions` map. Instead
the existing use of `onlyIncludeVersions` seems to do what's intended:
make dev builds faster. An `npm start` build takes about 10s both with
the old `disableVersioning` setting, or just relying on
`onlyIncludeVersions`.

This includes some cut-and-paste rearrangement too that makes the diff
harder to understand. The commits are individually reviewable.