Add VersionedDocsLink and use for evergreen links to "latest" docs page (#177)
This PR fixes #174 with a new `VersionedDocsLink` component that reads
the docusaurus config to determine what the latest stable version is.

It also adjusts the other dynamic link to the contribution guide to use
this. This link is changed from the latest RC versions (currently 2.20)
to the current `main` version (2.21), i.e. the absolute most up-to-date
version of the contribution guide, which seems most relevant to
contributors since they'll be starting in the `main` branch.

This has a few issues with the component due to
https://github.com/prettier/prettier/issues/12209, which forces us into
a very long line on the `_index.mdx` page.

I'm unclear if a broken link will be detected or not.