Override date for Pants 2.20 blog post (#191)
After merging #182, I realised that I'd forgotten to update date to the
publish date ("today") and thus it's quite wrong, because we took so
long to get 2.20 out the door. This PR fixes this by overriding the date
in the YAML front matter, so that it is displayed correctly in the post
metadata so users aren't confused about when 2.20 was released. This
doesn't change the filename because that would change the URL.