Fix dead changelog links in blog posts (#210)
Not sure if anything here is generated, but didn't look like it. I
couldn't find a perfect tag that followed the minor so I went with the
release one. Maybe you want them to target the latest patch for each,
but I also assume most/all of these will not be updated anymore anyways.