Escape $'s in reference default values  (#235)
This escapes `$` in default values to avoid them being treated as
template literal substitutions.

Default values end up in a template literal `` `something` ``. If the
string contains `${...}` literally, they'll be treated as a
substitution. This PR escapes all `$`s to avoid this.

This bug/oversight doesn't cause problems with any _current_ values, but
will be required in 2.23.0.dev4 due to the new default value for
`output_path` fields from
https://github.com/pantsbuild/pants/pull/21175.

Example of the effect (from
https://github.com/pantsbuild/pantsbuild.org/pull/234#discussion_r1690867967):

```diff
- default_repr={`'${spec_path_normalized}/${target_name_normalized}${file_suffix}'`}
+ default_repr={`'\${spec_path_normalized}/\${target_name_normalized}\${file_suffix}'`}
```