use friendly "semi-stable" urls for key versions (#256)
Currently all of the versioned docs have the explicit version in the
url. This has the virtue of being
[cool](https://www.w3.org/Provider/Style/URI) but has some awkwardness
in practice. For example if one links to
https://www.pantsbuild.org/2.18/docs/using-pants/using-pants-in-ci from
another webpage it will eventualy end up with the 'no longer maintained'
banner. It isn't super fun to update every internal/wiki link every 6
weeks for a release.

This change introduces some "semi-stable" urls for key versions:
 * `prerelease`: The latest prerelease
 * `dev`: aka `main` aka "trunk"
 * `stable`: The most recent stable release

Redirects ensure that the versioned urls also work.

As an example, right now:

`/stable/docs/using-pants/using-pants-in-ci` is the doc for `2.21` and
`/2.21//docs/using-pants/using-pants-in-ci` redirects to there. After
the next release, `/stable/docs/using-pants/using-pants-in-ci` will be
for `2.22` and `/2.21/docs/using-pants/using-pants-in-ci` will keep
pointing to the `2.21` content.

I think this is reasonable compromise of easy of use with permanence.

FWIW the Offical Docusaurus position seems to be to prefer server side
redirects https://github.com/facebook/docusaurus/issues/9049 but as far
as I can tell that isn't an option for GitHub Pages.

ref #221