Make sync_docs failures noisier by always posting a PR (#288)
This makes us more likely to notice problems in the automatic doc
syncing like #284 earlier, by surfacing the failures as PRs, not just
hidden in the actions UI.

The doc syncing usually happens automagically as part of the release
process, in the background. If it fails, this is only visible in two
ways:

- implicitly, in the lack of
<https://github.com/pantsbuild/pantsbuild.org/labels/automation%3Async-docs>-labelled
PR
- by noticing the failure on the actions page:
https://github.com/pantsbuild/pantsbuild.org/actions/workflows/sync_docs.yml

Both of these are easy to miss (very easy to miss that something that
should happen didn't happen, and the actions page is only rarely
visited!). There's no progress updates/reminder about the doc sync, nor
is there a notification on failure.

In practice, we missed failures on the 2.24.x branch for a few releases
(in #284).

This PR adds a "notification" about the failure, by posting the sync PR
unconditionally. The release manager thus gets a notification (about
their review request) and it's visible/easier for others to notice it
too, in the list of PRs. When there's a failure, this PR includes some
extra notes to highlight it.

Examples triggered from this branch: 

- successful run (unchanged):
https://github.com/pantsbuild/pantsbuild.org/pull/291
- unsuccessful run (new):
https://github.com/pantsbuild/pantsbuild.org/pull/290

---------

Co-authored-by: Josh Cannon <3956745+thejcannon@users.noreply.github.com>